+++ public/index.html
@@ -9,12 +9,12 @@
       name="description"
       content="Web site created using create-react-app"
     />
-    <meta data-hid="description" name="description" content="A simple app to airdrop your Matataki Token, Share to anywhere with the magical link.">
+    <meta data-hid="description" name="description" content="%META_DESC%">
     <!--  Meta for Twitter Card -->
     <meta data-hid="twitter:card" name="twitter:card" property="twitter:card" content="summary">
     <meta data-hid="twitter:site" name="twitter:site" property="twitter:site" content="@Andoromeda">
     <meta data-hid="twitter:title" name="twitter:title" property="twitter:title" content="Matataki Token AirDropper">
-    <meta data-hid="twitter:description" name="description" property="twitter:description" content="A simple app to airdrop your Matataki Token, Share to anywhere with the magical link.">
+    <meta data-hid="twitter:description" name="description" property="twitter:description" content="%META_DESC%">
     <!-- <meta data-hid="twitter:image" name="twitter:image" property="twitter:image" content="%PUBLIC_URL%/logo192.png"> -->
     <meta data-hid="twitter:image" name="twitter:image" property="twitter:image" content="%TOKEN_LOGO%">
     <!--  Meta for OpenGraph -->
@@ -23,7 +23,7 @@
     <meta data-hid="og:type" name="og:type" property="og:type" content="finance">
     <!-- <meta data-hid="og:image" name="og:image" property="og:image" content="%PUBLIC_URL%/logo192.png">  -->
     <meta data-hid="og:image" name="og:image" property="og:image" content="%TOKEN_LOGO%">
-    <meta data-hid="og:description" name="description" property="og:description" content="A simple app to airdrop your Matataki Token, Share to anywhere with the magical link.">
+    <meta data-hid="og:description" name="description" property="og:description" content="%META_DESC%">
     <!-- <link rel="apple-touch-icon" href="%PUBLIC_URL%/logo192.png" /> -->
     <link rel="apple-touch-icon" href="%TOKEN_LOGO%" />
     <link rel="manifest" href="%PUBLIC_URL%/manifest.json" />
